requirements.txt
runpod
transformers
torch
pillow 